refactoring(15): extract renderContext/renderFigures from ResultsPage
Move the two low-risk pure render blocks (experiment-context card,
figures grid) into resultSections.tsx as exported render-helper
functions, following the existing renderOrsChain/renderLegacyRepro
inline-call pattern (byte-identical bodies, same call semantics — no
component-boundary change). ResultsPage 1857->1744 lines.
