+++ ari-core/ari/viz/frontend/src/components/Results/ResultsPage.tsx
@@ -28,7 +28,7 @@ import { Card } from '../common/Card';
 import { Button } from '../common/Button';
 import { Badge } from '../common/Badge';
 import { PublishYamlEditor } from './PublishYamlEditor';
-import { renderOrsChain, renderLegacyRepro } from './resultSections';
+import { renderOrsChain, renderLegacyRepro, renderContext, renderFigures } from './resultSections';
 
 
 export function ResultsPage() {
@@ -1253,50 +1253,6 @@ export function ResultsPage() {
   };
 
   // Render experiment context
-  const renderContext = () => {
-    if (!summary) return null;
-    const sd = summary.science_data;
-    if (!sd || !(sd as any).experiment_context) return null;
-
-    const ctx = (sd as any).experiment_context as Record<string, unknown>;
-
-    return (
-      <Card style={{ marginBottom: 16 }}>
-        <div className="card-title">{t('exp_context')}</div>
-        <div>
-          {Object.entries(ctx).map(([k, v]) => {
-            const text = String(
-              typeof v === 'object'
-                ? JSON.stringify(v, null, 2)
-                : v,
-            );
-            return (
-              <div key={k} style={{ marginBottom: 12 }}>
-                <div style={{ color: 'var(--muted)', fontSize: '.75rem', marginBottom: 2, fontWeight: 600 }}>
-                  {k}
-                </div>
-                {text.length <= 500 ? (
-                  <div style={{ fontSize: '.8rem', whiteSpace: 'pre-wrap', wordBreak: 'break-word' }}>
-                    {text}
-                  </div>
-                ) : (
-                  <details>
-                    <summary style={{ cursor: 'pointer', color: 'var(--blue-light)', fontSize: '.75rem', listStyle: 'none', userSelect: 'none' }}>
-                      {'▶ Show detail'}
-                    </summary>
-                    <pre style={{ whiteSpace: 'pre-wrap', wordBreak: 'break-word', margin: '4px 0 0', fontSize: '.78rem', overflow: 'auto', maxHeight: 480 }}>
-                      {text}
-                    </pre>
-                  </details>
-                )}
-              </div>
-            );
-          })}
-        </div>
-      </Card>
-    );
-  };
-
   // Render Experiment Artifact Repository (EAR) section — issue #4
   const renderEAR = () => {
     if (earLoading) {
@@ -1694,75 +1650,6 @@ export function ResultsPage() {
   };
 
   // Render figures grid
-  const renderFigures = () => {
-    if (!summary) return null;
-    const fm = summary.figures_manifest as any;
-    if (!fm || !fm.figures) return null;
-
-    // plot-skill writes `figures` as a dict {name: path}; older runs may
-    // have stored a list [{path, caption, ...}]. Normalize to a list here
-    // so the grid works with both shapes.
-    const kinds = (fm.figure_kinds || {}) as Record<string, string>;
-    const snippets = (fm.latex_snippets || {}) as Record<string, string>;
-    const extractCaption = (snip: string): string => {
-      const m = snip.match(/\\caption\{([^}]+)\}/);
-      return m ? m[1] : '';
-    };
-    const figs: Array<{ name: string; path: string; caption: string; kind: string }> =
-      Array.isArray(fm.figures)
-        ? fm.figures.map((fig: any, idx: number) => ({
-            name: fig.name || `fig_${idx + 1}`,
-            path: fig.path || fig,
-            caption: fig.caption || '',
-            kind: fig.kind || fig.figure_kind || '',
-          }))
-        : Object.entries(fm.figures as Record<string, string>).map(([name, path]) => ({
-            name,
-            path: String(path),
-            caption: extractCaption(snippets[name] || ''),
-            kind: kinds[name] || '',
-          }));
-
-    if (!figs.length) return null;
-
-    return (
-      <Card style={{ marginBottom: 16 }}>
-        <div className="card-title">{'📈'} Figures</div>
-        <div className="grid-2">
-          {figs.map((fig, idx) => {
-            const { path, caption, kind } = fig;
-            const kindBadge: Record<string, string> = {
-              plot: 'Plot',
-              svg: 'Diagram',
-            };
-            return (
-              <div key={idx}>
-                <div style={{ display: 'flex', alignItems: 'center', gap: 6, marginBottom: 4 }}>
-                  {kind && kindBadge[kind] && (
-                    <Badge variant="blue">{kindBadge[kind]}</Badge>
-                  )}
-                  {caption && (
-                    <span style={{ fontSize: '.8rem', color: 'var(--muted)' }}>
-                      {caption}
-                    </span>
-                  )}
-                </div>
-                <img
-                  className="figure-img"
-                  src={`/codefile?path=${encodeURIComponent(path)}`}
-                  alt="figure"
-                  onError={(e) => {
-                    (e.target as HTMLImageElement).style.display = 'none';
-                  }}
-                />
-              </div>
-            );
-          })}
-        </div>
-      </Card>
-    );
-  };
-
   return (
     <div className="page active" style={{ display: 'block' }}>
       <h1>{t('results_title')}</h1>
@@ -1827,8 +1714,8 @@ export function ResultsPage() {
             {renderReviewScores()}
             {renderRepro()}
             {renderEAR()}
-            {renderContext()}
-            {renderFigures()}
+            {renderContext({ summary, t })}
+            {renderFigures({ summary })}
 
             {/* If no content at all */}
             {!summary.paper_tex &&
